+++ link-move/src/test/java/com/nhl/link/move/valueconverter/StringConverterTest.java
@@ -1,13 +1,11 @@
 package com.nhl.link.move.valueconverter;
 
-import com.nhl.link.move.LmRuntimeException;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class StringConverterTest {
 
@@ -40,52 +38,52 @@ public class StringConverterTest {
 
     @Test
     public void testConvert_Float() {
-        assertEquals("1", CONVERTER.convert(1f, -1));
+        assertEquals("1.0", CONVERTER.convert(1f, -1));
     }
 
     @Test
     public void testConvert_Float_NaN() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Float.NaN, -1));
+        assertEquals("NaN", CONVERTER.convert(Float.NaN, -1));
     }
 
     @Test
     public void testConvert_Float_PositiveInfinity() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Float.POSITIVE_INFINITY, -1));
+        assertEquals("Infinity", CONVERTER.convert(Float.POSITIVE_INFINITY, -1));
     }
 
     @Test
     public void testConvert_Float_NegativeInfinity() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Float.NEGATIVE_INFINITY, -1));
+        assertEquals("-Infinity", CONVERTER.convert(Float.NEGATIVE_INFINITY, -1));
     }
 
     @Test
     public void testConvert_Float_FractionalPart() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(1.1f, -1));
+        assertEquals("1.1", CONVERTER.convert(1.1f, -1));
     }
 
     @Test
     public void testConvert_Double() {
-        assertEquals("1", CONVERTER.convert(1d, -1));
+        assertEquals("1.0", CONVERTER.convert(1d, -1));
     }
 
     @Test
     public void testConvert_Double_NaN() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Double.NaN, -1));
+        assertEquals("NaN", CONVERTER.convert(Double.NaN, -1));
     }
 
     @Test
     public void testConvert_Double_PositiveInfinity() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Double.POSITIVE_INFINITY, -1));
+        assertEquals("Infinity", CONVERTER.convert(Double.POSITIVE_INFINITY, -1));
     }
 
     @Test
     public void testConvert_Double_NegativeInfinity() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(Double.NEGATIVE_INFINITY, -1));
+        assertEquals("-Infinity", CONVERTER.convert(Double.NEGATIVE_INFINITY, -1));
     }
 
     @Test
     public void testConvert_Double_FractionalPart() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(1.1d, -1));
+        assertEquals("1.1", CONVERTER.convert(1.1d, -1));
     }
 
     @Test
@@ -95,7 +93,7 @@ public class StringConverterTest {
 
     @Test
     public void testConvert_BigDecimal_Exception() {
-        assertThrows(LmRuntimeException.class, () -> CONVERTER.convert(BigDecimal.valueOf(1.1d), -1));
+        assertEquals("1.1", CONVERTER.convert(new BigDecimal("1.1"), -1));
     }
 
     @Test